math_qa_embedding/requirements.txt
pymilvus==2.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
flask==2.3.3 